refactor: split slack notification formatting helpers
