+++ api/services/notification_dispatcher.py
@@ -12,7 +12,6 @@ Block Kit support:
 
 import json
 import logging
-import re
 import threading
 import time
 import urllib.error
@@ -20,6 +19,7 @@ import urllib.request
 from typing import Dict, List, Optional
 
 from api.database import SessionLocal
+from api.services.notifications.slack_format import md_to_report_payload, md_to_slack_blocks
 
 logger = logging.getLogger(__name__)
 
@@ -32,235 +32,6 @@ _SLACK_RATE_LIMIT_SEC = 1.0
 _last_slack_send_time: float = 0.0
 # Lock guards _last_slack_send_time to prevent concurrent calls racing past the rate limiter
 _slack_lock = threading.Lock()
-
-
-# ── Block Kit helpers ──────────────────────────────────────────────────────────
-
-
-def _to_slack_mrkdwn(text: str) -> str:
-    """Convert markdown inline formatting to Slack mrkdwn syntax."""
-    # **bold** → *bold*
-    text = re.sub(r"\*\*([^*]+)\*\*", r"*\1*", text)
-    # *italic* (not already bold) → _italic_
-    text = re.sub(r"(?<!\*)\*([^*\n]+)\*(?!\*)", r"_\1_", text)
-    return text
-
-
-def _split_text(text: str, max_len: int = 2900) -> list[str]:
-    """Split text into chunks ≤ max_len, preferring newline boundaries."""
-    chunks: list[str] = []
-    while len(text) > max_len:
-        cut = text.rfind("\n", 0, max_len)
-        if cut == -1:
-            cut = max_len
-        chunks.append(text[:cut])
-        text = text[cut:].lstrip("\n")
-    if text:
-        chunks.append(text)
-    return chunks
-
-
-def md_to_slack_blocks(content: str) -> List[dict]:
-    """
-    Convert report markdown to Slack Block Kit blocks.
-
-    Mapping:
-      ## heading  → header block
-      ### heading → section *bold*
-      #### heading → context mrkdwn
-      | table |   → section fields (2-column grid, max 10 fields per section)
-      ---         → divider
-      > quote     → context block
-      text        → section mrkdwn (≤2900 chars, split if needed)
-    """
-    blocks: list[dict] = []
-    lines = content.split("\n")
-    text_buffer: list[str] = []
-    table_header: list[str] | None = None
-    table_rows: list[list[str]] = []
-
-    def flush_text() -> None:
-        nonlocal text_buffer
-        if not text_buffer:
-            return
-        joined = "\n".join(text_buffer).strip()
-        text_buffer = []
-        if not joined:
-            return
-        for chunk in _split_text(_to_slack_mrkdwn(joined)):
-            blocks.append({"type": "section", "text": {"type": "mrkdwn", "text": chunk}})
-
-    def flush_table() -> None:
-        nonlocal table_header, table_rows
-        if not table_header:
-            return
-
-        cols = len(table_header)
-        if cols == 2:
-            # 2-col indicator table → mrkdwn "*key*: value" per line
-            # (section fields forces 50/50 column split; text block respects content width)
-            lines_out: list[str] = []
-            for row in table_rows:
-                if len(row) < 2:
-                    continue
-                key = _to_slack_mrkdwn(row[0].strip())
-                val = _to_slack_mrkdwn(row[1].strip())
-                lines_out.append(f"*{key}*: {val or '—'}")
-            text = "\n".join(lines_out)
-            for chunk in _split_text(text):
-                blocks.append({"type": "section", "text": {"type": "mrkdwn", "text": chunk}})
-        else:
-            # Multi-col: header bold + rows as text
-            lines_out = [" | ".join(f"*{h.strip()}*" for h in table_header)]
-            for row in table_rows:
-                lines_out.append(" | ".join(_to_slack_mrkdwn(c.strip()) for c in row))
-            text = "\n".join(lines_out)
-            for chunk in _split_text(text):
-                blocks.append({"type": "section", "text": {"type": "mrkdwn", "text": chunk}})
-
-        table_header = None
-        table_rows.clear()
-
-    for line in lines:
-        stripped = line.strip()
-
-        if line.startswith("## "):
-            flush_text(); flush_table()
-            blocks.append({
-                "type": "header",
-                "text": {"type": "plain_text", "text": line[3:].strip(), "emoji": True},
-            })
-            continue
-
-        if line.startswith("### "):
-            flush_text(); flush_table()
-            blocks.append({
-                "type": "section",
-                "text": {"type": "mrkdwn", "text": f"*{_to_slack_mrkdwn(line[4:].strip())}*"},
-            })
-            continue
-
-        if line.startswith("#### "):
-            flush_text(); flush_table()
-            blocks.append({
-                "type": "context",
-                "elements": [{"type": "mrkdwn", "text": _to_slack_mrkdwn(line[5:].strip())}],
-            })
-            continue
-
-        if re.match(r"^-{3,}$", stripped):
-            flush_text(); flush_table()
-            blocks.append({"type": "divider"})
-            continue
-
-        if line.startswith("> "):
-            flush_text(); flush_table()
-            blocks.append({
-                "type": "context",
-                "elements": [{"type": "mrkdwn", "text": _to_slack_mrkdwn(line[2:])}],
-            })
-            continue
-
-        if line.startswith("|") and line.endswith("|"):
-            flush_text()
-            if re.match(r"^[\s|:-]+$", line):  # separator row
-                continue
-            cells = line[1:-1].split("|")
-            if table_header is None:
-                table_header = cells
-            else:
-                table_rows.append(cells)
-            continue
-
-        # Non-table line after a table
-        if table_header is not None:
-            flush_table()
-
-        if not stripped:
-            flush_text()
-            continue
-
-        text_buffer.append(line)
-
-    flush_text()
-    flush_table()
-    return blocks
-
-
-_SIGNAL_COLORS = {
-    "🔴": "#E74C3C",  # loss / danger
-    "🟡": "#F39C12",  # neutral / caution
-    "🟢": "#2ECC71",  # profit / good
-    "🎯": "#3498DB",  # action items
-}
-
-# Pattern: "#### N. 🔴 종목명 ..." or "### 🎯 핵심 액션 아이템"
-_STOCK_SECTION_RE = re.compile(r"^####\s+\d+\.\s+(🔴|🟡|🟢)")
-_ACTION_SECTION_RE = re.compile(r"^###\s+🎯")
-
-
-def md_to_report_payload(content: str) -> dict:
-    """
-    Convert report markdown to a full Slack payload with colored attachments.
-
-    Structure:
-      "blocks"      → header + macro section (no color)
-      "attachments" → one attachment per signal color (🔴/🟡/🟢) + one for action items (blue)
-                      Grouping avoids Slack's "Show less" collapse (triggered at >5 attachments).
-    """
-    lines = content.split("\n")
-
-    # Split into segments: (color | None, [lines])
-    segments: list[tuple[str | None, list[str]]] = []
-    current_color: str | None = None
-    current_lines: list[str] = []
-
-    for line in lines:
-        stock_m = _STOCK_SECTION_RE.match(line)
-        action_m = _ACTION_SECTION_RE.match(line)
-
-        if stock_m or action_m:
-            if current_lines:
-                segments.append((current_color, current_lines))
-            current_lines = [line]
-            if stock_m:
-                emoji = stock_m.group(1)
-                current_color = _SIGNAL_COLORS.get(emoji)
-            else:
-                current_color = _SIGNAL_COLORS["🎯"]
-        else:
-            current_lines.append(line)
-
-    if current_lines:
-        segments.append((current_color, current_lines))
-
-    # First segment (color=None) → main blocks
-    # Rest: group blocks by color to stay under Slack's ~5-attachment collapse threshold
-    main_blocks: list[dict] = []
-    # color → accumulated blocks (insertion-order preserved)
-    color_blocks: dict[str, list[dict]] = {}
-
-    for color, seg_lines in segments:
-        seg_blocks = md_to_slack_blocks("\n".join(seg_lines))
-        if color is None:
-            main_blocks.extend(seg_blocks)
-        else:
-            if color not in color_blocks:
-                color_blocks[color] = []
-            color_blocks[color].extend(seg_blocks)
-
-    # Build attachments: one attachment per color group, chunked at 50 blocks each
-    attachments: list[dict] = []
-    for color, blocks in color_blocks.items():
-        if not blocks:
-            continue
-        for i in range(0, len(blocks), 50):
-            attachments.append({
-                "color": color,
-                "blocks": blocks[i : i + 50],
-            })
-
-    return {"blocks": main_blocks, "attachments": attachments}
 
 
 def dispatch_report(
